requirements.txt
fastapi
uvicorn[standard]
typer[all]
httpx
python-dotenv
pydantic
pytest
pytest-asyncio
langchain>=0.2.0
langchain-groq
langgraph
langgraph-checkpoint-sqlite
langchain-huggingface
chromadb
sentence-transformers
langsmith
langchain_community
aiohttp
aiosqlite